app(upgrade) upgrade to Ionic 2.0.0-rc3
Upgrade to the latest Ionic release
 - Ionic 2.0.0-rc3
 - Angular 2.1.1
 - @ionic/app-scripts 0.0.45

Due to quite a lot of changes in the Ionic app-scripts package
the gulp task for live reloading has been moved back into the
src directory.

There is now hardcoding of the paths in the config files, which makes
it too difficult to control in the top level build package.

Will continue to review on each update to try and move this task
back out.